# Conflicts:
#	pom.xml
#	src/main/java/com/bookstrap/model/bean/AccountLabel.java
#	src/main/java/com/bookstrap/model/bean/Mail.java
#	src/main/java/com/bookstrap/model/bean/MailAccount.java
#	src/main/java/com/bookstrap/model/bean/MailAttachment.java
#	src/main/java/com/bookstrap/model/bean/MailCategory.java
#	src/main/java/com/bookstrap/model/bean/MailFolder.java
#	src/main/webapp/eddie/js/layoutfix.js

+++ src/main/java/com/bookstrap/harry/service/MemberDdetailService.java
@@ -5,10 +5,15 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.bookstrap.harry.bean.MemberDetails;
+import com.bookstrap.harry.bean.Members;
 import com.bookstrap.harry.dao.MemberDetailRepository;
 
 
@@ -33,15 +38,22 @@ public class MemberDdetailService {
 		return null;	
 				}
 	
-	public void deleteMemberDetailByID(Integer memberId) {
-		 memberDetailDao.deleteById(memberId);
+	public boolean deleteMemberDetailByID(Integer memberId) {
+		
+		if(memberId != null){
+		memberDetailDao.deleteById(memberId);
+		return true;
+		}
+		return false;
 	}
 
 	public MemberDetails insertMemberDetails(MemberDetails memberDatil) {
 		return memberDetailDao.save(memberDatil);
 	}
 	
-	public MemberDetails updateMemberDetails(Integer memberId, String memberName, Integer memberSex, String memberEmail, String memberCellPhone, String memberAddress, Date memberbirthday) {
+	public MemberDetails updateMemberDetails(Integer memberId, 
+			String memberLastName, String memberFirstName, Integer memberSex, String memberEmail, 
+			String memberCellPhone, String memberAddress, Date memberbirthday) {
 		Optional<MemberDetails> op = memberDetailDao.findById(memberId);
 	
 		if(op.isPresent()) {
@@ -49,8 +61,9 @@ public class MemberDdetailService {
 				memberDetail.setMemberAddress(memberAddress);
 				memberDetail.setMemberBirthday(memberbirthday);
 				memberDetail.setMemberEmail(memberEmail);
-				memberDetail.setMemberName(memberName);
-				memberDetail.setMemberPhone(memberName);
+				memberDetail.setMemberLastName(memberLastName);
+				memberDetail.setMemberFirstName(memberFirstName);
+				memberDetail.setMemberPhone(memberCellPhone);
 				memberDetail.setMemberSex(memberSex);
 				return memberDetail;
 		}
@@ -72,6 +85,12 @@ public class MemberDdetailService {
 			}
 			
 			return null;
+		}
+		
+		public Page<MemberDetails> getMemberByPage(Integer pageNumber){
+			Pageable pgb = PageRequest.of(pageNumber-1, 3, Sort.Direction.DESC, "memberId");
+				Page<MemberDetails> page = memberDetailDao.findAll(pgb);
+				return page;
 		}
